+++ CHANGELOG.md
@@ -1,8 +1,13 @@
+# 0.1.8
+- QBINOM (quantile in Binomial Distribution)
+- PBINOM (probability in Binomial Distribution)
+- RBINOM (quantiles in Binomial Distribution)
+
 # 0.1.7
 - QPOIS (quantile in Poisson Distribution)
 - PPOIS (probability in Poisson Distribution)
 - RPOIS (quantiles in Poisson Distribution)
-- 
+ 
 # 0.1.6
 - QF (quantile in F Distribution)
 - PF (probability in F Distribution)